Move showPasswordPrompt to PasswordRepromptService
