+++ src/abstractions/auth.service.ts
@@ -32,5 +32,4 @@ export abstract class AuthService {
     authingWithApiKey: () => boolean;
     authingWithSso: () => boolean;
     authingWithPassword: () => boolean;
-    showPasswordPrompt: () => Promise<boolean>;
 }
