+++ src/abstractions/passwordReprompt.service.ts
@@ -1,0 +1,3 @@
+export abstract class PasswordRepromptService {
+    showPasswordPrompt: () => Promise<boolean>;
+}
